Update tests with offsets calculated by librosa
Following the switch to librosa, the calculated offsets have changed slightly.  The changes are of the order of magnitude of one MFCC calculation window, which is similar to the accuracy of the system regardless.  The changes are therefore not considered problematic.

The details of the test audio given in the comments to the tests have also been updated following accurate measurements of the trimming using an audio editing tool.